+++ bugsnag-android-core/src/main/java/com/bugsnag/android/Client.java
@@ -2,9 +2,10 @@ package com.bugsnag.android;
 
 import static com.bugsnag.android.ContextExtensionsKt.getActivityManagerFrom;
 import static com.bugsnag.android.ContextExtensionsKt.getStorageManagerFrom;
-import static com.bugsnag.android.ImmutableConfigKt.sanitiseConfiguration;
 import static com.bugsnag.android.SeverityReason.REASON_HANDLED_EXCEPTION;
+import static com.bugsnag.android.internal.ImmutableConfigKt.sanitiseConfiguration;
 
+import com.bugsnag.android.internal.ImmutableConfig;
 import com.bugsnag.android.internal.StateObserver;
 
 import android.app.ActivityManager;
@@ -31,7 +32,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.RejectedExecutionException;
 
 /**
@@ -188,7 +188,7 @@ public class Client implements MetadataAware, CallbackAware, UserAware {
             sessionLifecycleCallback = new SessionLifecycleCallback(sessionTracker);
             application.registerActivityLifecycleCallbacks(sessionLifecycleCallback);
 
-            if (immutableConfig.shouldRecordBreadcrumbType(BreadcrumbType.STATE)) {
+            if (!immutableConfig.shouldDiscardBreadcrumb(BreadcrumbType.STATE)) {
                 this.activityBreadcrumbCollector = new ActivityBreadcrumbCollector(
                     new Function2<String, Map<String, ? extends Object>, Unit>() {
                         @SuppressWarnings("unchecked")
@@ -658,6 +658,9 @@ public class Client implements MetadataAware, CallbackAware, UserAware {
      */
     public void notify(@NonNull Throwable exc, @Nullable OnErrorCallback onError) {
         if (exc != null) {
+            if (immutableConfig.shouldDiscardError(exc)) {
+                return;
+            }
             SeverityReason severityReason = SeverityReason.newInstance(REASON_HANDLED_EXCEPTION);
             Metadata metadata = metadataState.getMetadata();
             Event event = new Event(exc, immutableConfig, severityReason, metadata, logger);
@@ -724,19 +727,6 @@ public class Client implements MetadataAware, CallbackAware, UserAware {
 
     void notifyInternal(@NonNull Event event,
                         @Nullable OnErrorCallback onError) {
-        String type = event.getImpl().getSeverityReasonType();
-        logger.d("Client#notifyInternal() - event captured by Client, type=" + type);
-        // Don't notify if this event class should be ignored
-        if (event.shouldDiscardClass()) {
-            logger.d("Skipping notification - should not notify for this class");
-            return;
-        }
-
-        if (!immutableConfig.shouldNotifyForReleaseStage()) {
-            logger.d("Skipping notification - should not notify for this release stage");
-            return;
-        }
-
         // set the redacted keys on the event as this
         // will not have been set for RN/Unity events
         Set<String> redactedKeys = metadataState.getMetadata().getRedactedKeys();
@@ -916,7 +906,7 @@ public class Client implements MetadataAware, CallbackAware, UserAware {
     void leaveAutoBreadcrumb(@NonNull String message,
                              @NonNull BreadcrumbType type,
                              @NonNull Map<String, Object> metadata) {
-        if (immutableConfig.shouldRecordBreadcrumbType(type)) {
+        if (!immutableConfig.shouldDiscardBreadcrumb(type)) {
             breadcrumbState.add(new Breadcrumb(message, type, metadata, new Date(), logger));
         }
     }