+++ bugsnag-android-core/src/test/java/com/bugsnag/android/ExceptionHandlerTest.kt
@@ -1,5 +1,6 @@
 package com.bugsnag.android
 
+import com.bugsnag.android.internal.ImmutableConfig
 import org.junit.After
 import org.junit.Assert.assertSame
 import org.junit.Assert.assertTrue
@@ -9,6 +10,7 @@ import org.junit.runner.RunWith
 import org.mockito.ArgumentMatchers.any
 import org.mockito.ArgumentMatchers.eq
 import org.mockito.Mock
+import org.mockito.Mockito.`when`
 import org.mockito.Mockito.times
 import org.mockito.Mockito.verify
 import org.mockito.junit.MockitoJUnitRunner
@@ -20,11 +22,15 @@ internal class ExceptionHandlerTest {
     @Mock
     lateinit var client: Client
 
+    @Mock
+    lateinit var cfg: ImmutableConfig
+
     var originalHandler: Thread.UncaughtExceptionHandler? = null
 
     @Before
     fun setUp() {
         originalHandler = Thread.getDefaultUncaughtExceptionHandler()
+        `when`(client.config).thenReturn(cfg)
     }
 
     @After
@@ -66,5 +72,20 @@ internal class ExceptionHandlerTest {
         val thread = Thread.currentThread()
         exceptionHandler.uncaughtException(thread, RuntimeException("Whoops"))
         assertTrue(propagated)
+    }
+
+    @Test
+    fun uncaughtExceptionOutsideReleaseStages() {
+        val exceptionHandler = ExceptionHandler(client, NoopLogger)
+        val thread = Thread.currentThread()
+        val exc = RuntimeException("Whoops")
+        `when`(cfg.shouldDiscardError(exc)).thenReturn(true)
+        exceptionHandler.uncaughtException(thread, exc)
+        verify(client, times(0)).notifyUnhandledException(
+            eq(exc),
+            any(),
+            eq(SeverityReason.REASON_UNHANDLED_EXCEPTION),
+            eq(null)
+        )
     }
 }