+++ features/fixtures/mazerunner/jvm-scenarios/src/main/java/com/bugsnag/android/mazerunner/scenarios/NullReleaseStageScenario.kt
@@ -23,6 +23,5 @@ internal class NullReleaseStageScenario(
         Bugsnag.notify(generateException())
     }
 
-    override fun getInterceptedLogMessages() =
-        listOf("Skipping notification - should not notify for this release stage")
+    override fun getInterceptedLogMessages() = listOf("Bugsnag loaded")
 }